Phase 2 follow-ups: plan -I drift, type consistency, clangd -xc
Code review of Phase 2 flagged one Important issue plus two small
Minor items worth fixing before Phase 3 starts.

Important:
  plan-1-host-pipeline.md carried stale per-module -I flags in six
  per-task verify-compile snippets (lines 1132/1205/1301/1385/1461/
  1559/2086) and in Phase 10's cppcheck workflow (line 5751). After
  the earlier path-qualified-includes remediation, those snippets
  would fail when a Phase 3+ subagent tried to run them ("file not
  found" on path-qualified includes like render/arena.h). All seven
  sites replaced with `-I.` (repo root) to match the headers' and
  the Makefile's convention.

Minor:
  - test/fake_syscalls.h: new_handle_fail_after changes from short
    to int for consistency with set_handle_size_fail_after (both
    serve identical "-1 = never, N = fail on call N+1" role).
  - mdparse/mdparse.h: new docstring on MdParseSink explains the
    callback return convention (0 = continue, non-zero = abort
    with kMdParseErrSinkAbort). Was previously discoverable only
    by reading mdparse_run's impl — now documented at the type.
  - .clangd: -xc added to CompileFlags so clangd treats .h files
    as C instead of defaulting to Objective-C++ on macOS. Fixes
    "Invalid argument '-std=c89' not allowed with 'Objective-C++'"
    IDE diagnostics.

Deferred to Phase 3+ review: g_current singleton docstring
expansion (safe as-is; will surface if tests misuse it), RenderError
enum documentation (Phase 8 will use it), SrcPaneParams Rect-ordering
comment, Block.text NUL-termination note prominence.

+++ test/fake_syscalls.h
@@ -22,7 +22,7 @@ typedef struct FakeSyscalls {
     short         get_eof_result;
     long          get_eof_size;                /* value returned via *out_size */
     short         mem_error_result;
-    short         new_handle_fail_after;       /* -1 = never */
+    int           new_handle_fail_after;       /* -1 = never */
     short         gestalt_result;
     long          gestalt_response;            /* value returned via *out_response */
